rename models; fix seed to clear DB
