+++ controllers/users.js
@@ -1,5 +1,5 @@
 // Require resource's model(s).
-var User = require("../models/users");
+var User = require("../models/user");
 
 module.exports = {
   index: function (req, res, next) {
